move model debugging utils to dedicated file
